获取在线用户接口中添加total数据总数

+++ CCB-BackendServer/pojo/src/main/java/com/ecnu/vo/OnlineCounselorVO.java
@@ -16,5 +16,5 @@ public class OnlineCounselorVO implements Serializable {
 
     List<OnlineCounselor> onlineCounselors;
 
-    Integer total;
+    Long total;
 }
