+++ CCB-BackendServer/pojo/src/main/java/com/ecnu/vo/OnlineSupervisorVO.java
@@ -16,5 +16,5 @@ public class OnlineSupervisorVO implements Serializable {
 
     List<OnlineSupervisor> onlineSupervisors;
 
-    Integer total;
+    Long total;
 }
